Added first implementation for choosing on graph
There is a lot of hacks and bad code, so in future we have to take some
time for cleaning and improving our code rules, styles and more. Also
we can move from less to sass (scss) and use ES6 inside browser with
Babel. Also we can implement here flowtype or typescript.